server_state
Divede server state from response_manager.
